Final touch: move GenericTypeNode type from ClassStringTypeMapper to dedicated GenericClassStringTypeMapper
